+++ corehq/apps/app_manager/tests/test_suite_case_tiles.py
@@ -13,59 +13,58 @@ from corehq.apps.app_manager.suite_xml.features.case_tiles import CaseTileTempla
 from corehq.apps.app_manager.tests.app_factory import AppFactory
 from corehq.apps.app_manager.tests.util import (
     SuiteMixin,
-    TestXmlMixin,
     patch_get_xform_resource_overrides,
 )
 from corehq.util.test_utils import flag_enabled
 
 
+def add_columns_for_case_details(_module):
+    _module.case_details.short.columns = [
+        DetailColumn(
+            header={'en': 'a'},
+            model='case',
+            field='a',
+            format='plain',
+            case_tile_field='header'
+        ),
+        DetailColumn(
+            header={'en': 'b'},
+            model='case',
+            field='b',
+            format='plain',
+            case_tile_field='top_left'
+        ),
+        DetailColumn(
+            header={'en': 'c'},
+            model='case',
+            field='c',
+            format='enum',
+            enum=[
+                MappingItem(key='male', value={'en': 'Male'}),
+                MappingItem(key='female', value={'en': 'Female'}),
+            ],
+            case_tile_field='sex'
+        ),
+        DetailColumn(
+            header={'en': 'd'},
+            model='case',
+            field='d',
+            format='address',
+            case_tile_field='bottom_left'
+        ),
+        DetailColumn(
+            header={'en': 'e'},
+            model='case',
+            field='e',
+            format='date',
+            case_tile_field='date'
+        ),
+    ]
+
+
 @patch_get_xform_resource_overrides()
 class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
     file_path = ('data', 'suite')
-
-    @staticmethod
-    def _add_columns_for_case_details(_module):
-        _module.case_details.short.columns = [
-            DetailColumn(
-                header={'en': 'a'},
-                model='case',
-                field='a',
-                format='plain',
-                case_tile_field='header'
-            ),
-            DetailColumn(
-                header={'en': 'b'},
-                model='case',
-                field='b',
-                format='plain',
-                case_tile_field='top_left'
-            ),
-            DetailColumn(
-                header={'en': 'c'},
-                model='case',
-                field='c',
-                format='enum',
-                enum=[
-                    MappingItem(key='male', value={'en': 'Male'}),
-                    MappingItem(key='female', value={'en': 'Female'}),
-                ],
-                case_tile_field='sex'
-            ),
-            DetailColumn(
-                header={'en': 'd'},
-                model='case',
-                field='d',
-                format='address',
-                case_tile_field='bottom_left'
-            ),
-            DetailColumn(
-                header={'en': 'e'},
-                model='case',
-                field='e',
-                format='date',
-                case_tile_field='date'
-            ),
-        ]
 
     def ensure_module_session_datum_xml(self, factory, detail_inline_attr, detail_persistent_attr):
         suite = factory.app.create_suite()
@@ -100,7 +99,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         module.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
         module.case_details.short.persist_tile_on_forms = True
         module.case_details.short.pull_down_tile = True
-        self._add_columns_for_case_details(module)
+        add_columns_for_case_details(module)
 
         form = app.new_form(0, "Untitled Form", None)
         form.xmlns = 'http://id_m0-f0'
@@ -119,7 +118,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         module.case_type = 'patient'
         module.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
         module.case_details.short.use_case_tiles = True
-        self._add_columns_for_case_details(module)
+        add_columns_for_case_details(module)
 
         form = app.new_form(0, "Untitled Form", None)
         form.xmlns = 'http://id_m0-f0'
@@ -136,7 +135,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         module0, form0 = factory.new_advanced_module("m0", "person")
         factory.form_requires_case(form0, "person")
         module0.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
-        self._add_columns_for_case_details(module0)
+        add_columns_for_case_details(module0)
 
         module1, form1 = factory.new_advanced_module("m1", "person")
         factory.form_requires_case(form1, "person")
@@ -165,7 +164,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         # persists case tiles and detail inline from another module
         module1.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
         module1.case_details.short.persist_tile_on_forms = True
-        self._add_columns_for_case_details(module1)
+        add_columns_for_case_details(module1)
         self.ensure_module_session_datum_xml(factory, 'detail-inline="m0_case_long"',
                                              'detail-persistent="m0_case_short"')
 
@@ -186,7 +185,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         module0, form0 = factory.new_advanced_module("m0", "person")
         factory.form_requires_case(form0, "person")
         module0.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
-        self._add_columns_for_case_details(module0)
+        add_columns_for_case_details(module0)
 
         module1, form1 = factory.new_advanced_module("m1", "person")
         factory.form_requires_case(form1, "person")
@@ -210,7 +209,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         # persists case tiles from another module
         module1.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
         module1.case_details.short.persist_tile_on_forms = True
-        self._add_columns_for_case_details(module1)
+        add_columns_for_case_details(module1)
         self.ensure_module_session_datum_xml(factory, '', 'detail-persistent="m0_case_short"')
 
         # set to use case tile from a module that does not support case tiles anymore
@@ -392,7 +391,7 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
         module = app.add_module(Module.new_module('Untitled Module', None))
         module.case_type = 'patient'
         module.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
-        self._add_columns_for_case_details(module)
+        add_columns_for_case_details(module)
 
         module.search_config = CaseSearch(
             properties=[
@@ -453,63 +452,4 @@ class SuiteCaseTilesTest(SimpleTestCase, SuiteMixin):
             app.create_suite(),
             # action[1] is the reg from case list action hard-coded into the default template
             "detail[@id='m0_search_short']/action[2]",
-        )
-
-    def test_case_tiles_with_grouping(self, *args):
-        app = Application.new_app('domain', 'Untitled Application')
-
-        module = app.add_module(Module.new_module('Untitled Module', None))
-        module.case_type = 'patient'
-        module.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
-        self._add_columns_for_case_details(module)
-        module.case_details.short.case_tile_group = CaseTileGroupConfig(xpath_function="./index/parent")
-
-        module.assign_references()
-
-        form = app.new_form(0, "Untitled Form", None)
-        form.xmlns = 'http://id_m0-f0'
-        form.requires = 'case'
-
-        self.assertXmlPartialEqual(
-            """
-            <partial>
-               <group function="./index/parent" grid-header-rows="1"/>
-            </partial>
-            """,
-            app.create_suite(),
-            "detail[@id='m0_case_short']/group",
-        )
-
-    def test_entry_datum_with_case_tile_grouping(self, *args):
-        app = Application.new_app('domain', 'Untitled Application')
-
-        module = app.add_module(Module.new_module('Untitled Module', None))
-        module.case_type = 'child'
-        module.case_details.short.case_tile_template = CaseTileTemplates.PERSON_SIMPLE.value
-        self._add_columns_for_case_details(module)
-        module.case_details.short.case_tile_group = CaseTileGroupConfig(xpath_function="./index/parent")
-
-        module.assign_references()
-
-        form = app.new_form(0, "Untitled Form", None)
-        form.xmlns = 'http://id_m0-f0'
-        form.requires = 'case'
-
-        suite = app.create_suite()
-        self.assertXmlPartialEqual(
-            """
-            <partial>
-              <session>
-                <datum 
-                    detail-confirm="m0_case_long" 
-                    detail-select="m0_case_short" 
-                    id="case_id_child" 
-                    nodeset="instance('casedb')/casedb/case[@case_type='child'][@status='open']" 
-                    value="./@case_id"/>
-                <datum function="./index/parent" id="case_id"/>
-              </session>
-            </partial>
-            """,
-            suite,
-            "entry[1]/session",
         )
